+++ examples/file_body.hpp
@@ -13,6 +13,7 @@
 #include <boost/asio/buffer.hpp>
 #include <boost/assert.hpp>
 #include <boost/filesystem.hpp>
+#include <boost/optional.hpp>
 #include <cstdio>
 #include <cstdint>
 
@@ -33,6 +34,12 @@ struct file_body
         std::size_t buf_len_;
 
     public:
+        using is_deferred = std::true_type;
+
+        using const_buffers_type =
+            boost::asio::const_buffers_1;
+
+        writer(writer&&) = default;
         writer(writer const&) = delete;
         writer& operator=(writer const&) = delete;
 
@@ -87,6 +94,27 @@ struct file_body
             offset_ += nread;
             wf(boost::asio::buffer(buf_, nread));
             return offset_ >= size_;
+        }
+
+        boost::optional<std::pair<const_buffers_type, bool>>
+        read(error_code& ec)
+        {
+            if(size_ - offset_ < sizeof(buf_))
+                buf_len_ = static_cast<std::size_t>(
+                    size_ - offset_);
+            else
+                buf_len_ = sizeof(buf_);
+            auto const nread = fread(
+                buf_, 1, sizeof(buf_), file_);
+            if(ferror(file_))
+            {
+                ec = error_code(errno, system_category());
+                return boost::none;
+            }
+            BOOST_ASSERT(nread != 0);
+            offset_ += nread;
+            return {{const_buffers_type{buf_, nread},
+                offset_ >= size_}};
         }
     };
 };