+++ test/http/write.cpp
@@ -8,13 +8,15 @@
 // Test that header file is self-contained.
 #include <beast/http/write.hpp>
 
+#include <beast/http/buffer_body.hpp>
 #include <beast/http/fields.hpp>
 #include <beast/http/message.hpp>
+#include <beast/http/read.hpp>
 #include <beast/http/string_body.hpp>
-#include <beast/http/write.hpp>
 #include <beast/core/error.hpp>
 #include <beast/core/multi_buffer.hpp>
 #include <beast/test/fail_stream.hpp>
+#include <beast/test/string_istream.hpp>
 #include <beast/test/string_ostream.hpp>
 #include <beast/test/yield_to.hpp>
 #include <beast/unit_test/suite.hpp>
@@ -39,6 +41,11 @@ public:
             value_type const& body_;
 
         public:
+            using is_deferred = std::false_type;
+
+            using const_buffers_type =
+                boost::asio::const_buffers_1;
+
             template<bool isRequest, class Allocator>
             explicit
             writer(message<isRequest, unsized_body, Allocator> const& msg) noexcept
@@ -59,6 +66,133 @@ public:
                 wf(boost::asio::buffer(body_));
                 return true;
             }
+            
+            boost::optional<std::pair<const_buffers_type, bool>>
+            read(error_code&)
+            {
+                return {{const_buffers_type{
+                    body_.data(), body_.size()}, false}};
+            }
+        };
+    };
+
+    template<
+        bool isDeferred,
+        bool isSplit,
+        bool isFinalEmpty
+    >
+    struct test_body
+    {
+        using value_type = std::string;
+
+        class writer
+        {
+            int step_ = 0;
+            value_type const& body_;
+
+        public:
+            using is_deferred =
+                std::integral_constant<bool, isDeferred>;
+
+            using const_buffers_type =
+                boost::asio::const_buffers_1;
+
+            template<bool isRequest, class Fields>
+            explicit
+            writer(message<isRequest, test_body,
+                    Fields> const& msg)
+                : body_(msg.body)
+            {
+            }
+
+            void
+            init(error_code&)
+            {
+            }
+
+            boost::optional<std::pair<const_buffers_type, bool>>
+            read(error_code&)
+            {
+                return read(
+                    std::integral_constant<bool, isSplit>{},
+                    std::integral_constant<bool, isFinalEmpty>{});
+            }
+
+        private:
+            boost::optional<std::pair<const_buffers_type, bool>>
+            read(
+                std::false_type,    // isSplit
+                std::false_type)    // isFinalEmpty
+            {
+                using boost::asio::buffer;
+                if(body_.empty())
+                    return boost::none;
+                return {{buffer(body_.data(), body_.size()), false}};
+            }
+
+            boost::optional<std::pair<const_buffers_type, bool>>
+            read(
+                std::false_type,    // isSplit
+                std::true_type)     // isFinalEmpty
+            {
+                using boost::asio::buffer;
+                if(body_.empty())
+                    return boost::none;
+                switch(step_)
+                {
+                case 0:
+                    step_ = 1;
+                    return {{buffer(
+                        body_.data(), body_.size()), true}};
+                default:
+                    return boost::none;
+                }
+            }
+
+            boost::optional<std::pair<const_buffers_type, bool>>
+            read(
+                std::true_type,     // isSplit
+                std::false_type)    // isFinalEmpty
+            {
+                using boost::asio::buffer;
+                auto const n = (body_.size() + 1) / 2;
+                switch(step_)
+                {
+                case 0:
+                    if(n == 0)
+                        return boost::none;
+                    step_ = 1;
+                    return {{buffer(body_.data(), n),
+                        body_.size() > 1}};
+                default:
+                    return {{buffer(body_.data() + n,
+                        body_.size() - n), false}};
+                }
+            }
+
+            boost::optional<std::pair<const_buffers_type, bool>>
+            read(
+                std::true_type,     // isSplit
+                std::true_type)     // isFinalEmpty
+            {
+                using boost::asio::buffer;
+                auto const n = (body_.size() + 1) / 2;
+                switch(step_)
+                {
+                case 0:
+                    if(n == 0)
+                        return boost::none;
+                    step_ = body_.size() > 1 ? 1 : 2;
+                    return {{buffer(body_.data(), n), true}};
+                case 1:
+                    BOOST_ASSERT(body_.size() > 1);
+                    step_ = 2;
+                    return {{buffer(body_.data() + n,
+                        body_.size() - n), true}};
+                default:
+                    return boost::none;
+                }
+            }
         };
     };
 
@@ -72,20 +206,12 @@ public:
 
             std::string s_;
             test::fail_counter& fc_;
-            boost::asio::io_service& ios_;
 
         public:
-            value_type(test::fail_counter& fc,
-                    boost::asio::io_service& ios)
+            explicit
+            value_type(test::fail_counter& fc)
                 : fc_(fc)
-                , ios_(ios)
             {
-            }
-
-            boost::asio::io_service&
-            get_io_service() const
-            {
-                return ios_;
             }
 
             value_type&
@@ -102,6 +228,11 @@ public:
             value_type const& body_;
 
         public:
+            using is_deferred = std::false_type;
+
+            using const_buffers_type =
+                boost::asio::const_buffers_1;
+
             template<bool isRequest, class Allocator>
             explicit
             writer(message<isRequest, fail_body, Allocator> const& msg) noexcept
@@ -127,8 +258,31 @@ public:
                 ++n_;
                 return n_ == body_.s_.size();
             }
+
+            boost::optional<std::pair<const_buffers_type, bool>>
+            read(error_code& ec)
+            {
+                if(body_.fc_.fail(ec))
+                    return boost::none;
+                if(n_ >= body_.s_.size())
+                    return boost::none;
+                return {{const_buffers_type{
+                    body_.s_.data() + n_++, 1}, true}};
+            }
         };
     };
+
+    template<bool isRequest>
+    bool
+    equal_body(string_view sv, string_view body)
+    {
+        test::string_istream si{
+            get_io_service(), sv.to_string()};
+        message<isRequest, string_body, fields> m;
+        multi_buffer b;
+        read(si, b, m);
+        return m.body == body;
+    }
 
     template<bool isRequest, class Body, class Fields>
     std::string
@@ -232,9 +386,7 @@ public:
             test::fail_counter fc(n);
             test::fail_stream<
                 test::string_ostream> fs(fc, ios_);
-            message<true, fail_body, fields> m(
-                std::piecewise_construct,
-                    std::forward_as_tuple(fc, ios_));
+            message<true, fail_body, fields> m{fc};
             m.method("GET");
             m.target("/");
             m.version = 10;
@@ -265,9 +417,7 @@ public:
             test::fail_counter fc(n);
             test::fail_stream<
                 test::string_ostream> fs(fc, ios_);
-            message<true, fail_body, fields> m(
-                std::piecewise_construct,
-                    std::forward_as_tuple(fc, ios_));
+            message<true, fail_body, fields> m{fc};
             m.method("GET");
             m.target("/");
             m.version = 10;
@@ -300,9 +450,7 @@ public:
             test::fail_counter fc(n);
             test::fail_stream<
                 test::string_ostream> fs(fc, ios_);
-            message<true, fail_body, fields> m(
-                std::piecewise_construct,
-                    std::forward_as_tuple(fc, ios_));
+            message<true, fail_body, fields> m{fc};
             m.method("GET");
             m.target("/");
             m.version = 10;
@@ -335,9 +483,7 @@ public:
             test::fail_counter fc(n);
             test::fail_stream<
                 test::string_ostream> fs(fc, ios_);
-            message<true, fail_body, fields> m(
-                std::piecewise_construct,
-                    std::forward_as_tuple(fc, ios_));
+            message<true, fail_body, fields> m{fc};
             m.method("GET");
             m.target("/");
             m.version = 10;
@@ -365,9 +511,7 @@ public:
             test::fail_counter fc(n);
             test::fail_stream<
                 test::string_ostream> fs(fc, ios_);
-            message<true, fail_body, fields> m(
-                std::piecewise_construct,
-                    std::forward_as_tuple(fc, ios_));
+            message<true, fail_body, fields> m{fc};
             m.method("GET");
             m.target("/");
             m.version = 10;
@@ -664,6 +808,179 @@ public:
         }
     }
 
+    template<class Stream,
+        bool isRequest, class Body, class Fields>
+    static
+    void
+    do_write(Stream& stream,
+        message<isRequest, Body, Fields> const& m,
+            error_code& ec)
+    {
+        auto ws = make_write_stream(m);
+        for(;;)
+        {
+            ws.write_some(stream, ec);
+            if(ec)
+                return;
+            if(ws.is_done())
+                break;
+        }
+    }
+
+    template<class Stream,
+        bool isRequest, class Body, class Fields>
+    static
+    void
+    do_async_write(Stream& stream,
+        message<isRequest, Body, Fields> const& m,
+            error_code& ec, yield_context yield)
+    {
+        auto ws = make_write_stream(m);
+        for(;;)
+        {
+            ws.async_write_some(stream, yield[ec]);
+            if(ec)
+                return;
+            if(ws.is_done())
+                break;
+        }
+    }
+
+    template<class Body>
+    void
+    testWriteStream(boost::asio::yield_context yield)
+    {
+        message<false, Body, fields> m;
+        m.version = 11;
+        m.status = 200;
+        m.reason("OK");
+        m.fields.insert("Server", "test");
+        m.body = "Hello, world!\n";
+        {
+            std::string const result =
+                "HTTP/1.1 200 OK\r\n"
+                "Server: test\r\n"
+                "\r\n"
+                "Hello, world!\n";
+            {
+                error_code ec;
+                test::string_ostream so{get_io_service(), 3};
+                do_write(so, m, ec);
+                BEAST_EXPECT(so.str == result);
+                BEAST_EXPECT(equal_body<false>(so.str, m.body));
+            }
+            {
+                error_code ec;
+                test::string_ostream so{get_io_service(), 3};
+                do_async_write(so, m, ec, yield);
+                BEAST_EXPECT(so.str == result);
+                BEAST_EXPECT(equal_body<false>(so.str, m.body));
+            }
+        }
+        {
+            m.fields.insert("Transfer-Encoding", "chunked");
+            {
+                error_code ec;
+                test::string_ostream so{get_io_service(), 3};
+                do_write(so, m, ec);
+                BEAST_EXPECT(equal_body<false>(so.str, m.body));
+            }
+            {
+                error_code ec;
+                test::string_ostream so{get_io_service(), 3};
+                do_async_write(so, m, ec, yield);
+                BEAST_EXPECT(equal_body<false>(so.str, m.body));
+            }
+        }
+    }
+
+    /** Execute a child process and return the output as an HTTP response.
+
+        @param input A stream to read the child process output from.
+
+        @param output A stream to write the HTTP response to.
+    */
+    template<class SyncReadStream, class SyncWriteStream>
+    void
+    cgi_process(SyncReadStream& input, SyncWriteStream& output, error_code& ec)
+    {
+        multi_buffer b;
+        message<false, buffer_body<true,
+            typename multi_buffer::const_buffers_type>, fields> m;
+        m.status = 200;
+        m.version = 11;
+        m.fields.insert("Server", "cgi-process");
+        m.fields.insert("Transfer-Encoding", "chunked");
+        m.body.first = boost::none;
+        m.body.second = true;
+
+        auto w = make_write_stream(m);
+
+        // send the header first, so the
+        // other end gets it right away
+        for(;;)
+        {
+            w.write_some(output, ec);
+            if(ec == error::need_more)
+            {
+                ec = {};
+                break;
+            }
+            if(ec)
+                return;
+        }
+
+        // send the body
+        for(;;)
+        {
+            // read from input
+            auto bytes_transferred =
+                input.read_some(b.prepare(1024), ec);
+            if(ec == boost::asio::error::eof)
+            {
+                BOOST_ASSERT(bytes_transferred == 0);
+                ec = {};
+                m.body = {boost::none, false};
+            }
+            else
+            {
+                if(ec)
+                    return;
+                b.commit(bytes_transferred);
+                m.body = {b.data(), true};
+            }
+            
+            // write to output
+            for(;;)
+            {
+                w.write_some(output, ec);
+                if(ec == error::need_more)
+                {
+                    ec = {};
+                    break;
+                }
+                if(ec)
+                    return;
+                if(w.is_done())
+                    goto is_done;
+            }
+            b.consume(b.size());
+        }
+    is_done:
+        ;
+    }
+
+    void
+    testCgiRelay()
+    {
+        error_code ec;
+        std::string const body = "Hello, world!\n";
+        test::string_ostream so{get_io_service(), 3};
+        test::string_istream si{get_io_service(), body, 6};
+        cgi_process(si, so, ec);
+        BEAST_EXPECT(equal_body<false>(so.str, body));
+    }
+
     void run() override
     {
         yield_to(&write_test::testAsyncWriteHeaders, this);
@@ -673,6 +990,19 @@ public:
         test_std_ostream();
         testOstream();
         testIoService();
+        testCgiRelay();
+        yield_to(
+            [&](boost::asio::yield_context yield)
+            {
+                testWriteStream<test_body<false, false, false>>(yield);
+                testWriteStream<test_body<false, false,  true>>(yield);
+                testWriteStream<test_body<false,  true, false>>(yield);
+                testWriteStream<test_body<false,  true,  true>>(yield);
+                testWriteStream<test_body< true, false, false>>(yield);
+                testWriteStream<test_body< true, false,  true>>(yield);
+                testWriteStream<test_body< true,  true, false>>(yield);
+                testWriteStream<test_body< true,  true,  true>>(yield);
+            });
     }
 };